+++ tools/megatest.py
@@ -137,7 +137,6 @@ def buildpython(workdir, pyver, ucs, logfilename):
 # Default versions we support
 PYVERS=(
     '3.1.1',
-    '3.0.1',
     '2.6.3',
     '2.5.4',
     '2.4.6',